--- frontend/src/Admin/Admin.css
@@ -1,108 +1,0 @@
-.admin-container {
-    display: flex;
-    width: 100vw;
-    height: 100vh;
-    background: #fff;
-    overflow: hidden;
-    color: black;
-}
-/* alterar o nome e local pois esta a ser usado para outras pages  */
-.main-content {
-    margin-top: 1.8%;
-    margin-left: 300px;
-    padding: 20px;
-    flex-grow: 1;
-}
-
-.heading-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-  
-.heading {
-    font-size: 24px;
-    font-weight: bold;
-    margin-top: 1%;
-}
-  
-.dropdown-container {
-    position: relative;
-}
-  
-.order-button {
-    background-color: #28a745;
-    border: none;
-    color: white;
-    padding: 8px 16px;
-    font-size: 14px;
-    cursor: pointer;
-    border-radius: 5px;
-    margin-right: 150px;
-}
-  
-.order-button:hover {
-    background-color: #218838;
-}
-  
-.dropdown-menu {
-    position: absolute;
-    top: 100%;
-    right: 0;
-    background-color: white;
-    border: 1px solid #ccc;
-    box-shadow: 0px 8px 16px rgba(0, 0, 0, 0.2);
-    z-index: 1000;
-    min-width: 150px;
-    margin-top: 5px;
-}
-  
-.dropdown-item {
-    padding: 10px;
-    cursor: pointer;
-    font-size: 14px;
-}
-  
-.dropdown-item:hover {
-    background-color: #f1f1f1;
-}
-  
-.dropdown-item:active {
-    background-color: #ddd;
-}
-
-/* Cards */
-
-.algorithm-cards {
-    display: flex;
-    justify-content: flex-start;
-    gap: 40px;
-    margin-top: 20px;
-  }
-  
-  .algorithm-card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 220px;
-    height: 120px;
-    border-radius: 10px;
-    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
-  }
-  
-  .add-button {
-    background: transparent;
-    border: 2px solid black;
-    color: black;
-    font-size: 16px;
-    cursor: pointer;
-    margin-top: 20px;
-    text-align: left;
-  }
-  
-  .purple { background-color: #e0d4fc; }
-  .green { background-color: #d4fcd4; }
-  .yellow { background-color: #fcf4d4; }
-  
-
