+++ frontend/src/Admin/Admin.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import Sidebar from "../components/Sidebar"; 
 import AlgorithmCard from "../components/admin/AlgorithmCard";
-import "./Admin.css";
+import "../styles/Admin.css";
 
 const Admin = () => {
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
