--- frontend/src/Manager/Calendar.css
@@ -1,4 +1,0 @@
-.main-content {
-    margin-right: 0 !important;
-}
-
