--- frontend/src/Manager/manager.css
@@ -1,16 +1,0 @@
-.btn {
-    display: inline-block;
-    padding: 10px 20px;
-    background-color: #007bff; 
-    color: white;
-    text-align: center;
-    text-decoration: none;
-    border-radius: 5px;
-    font-size: 16px;
-    transition: background-color 0.3s ease;
-    margin: 20px;
-}
-  
-.btn:hover {
-    background-color: #0056b3;
-}
