--- frontend/src/components/Sidebar.css
@@ -1,132 +1,0 @@
-.sidebar {
-  height: 100%;
-  width: 250px;
-  background-color: #e3f2fd;
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  position: fixed;
-  top: 0;
-  left: 0;
-  overflow: hidden;
-}
-
-.logo {
-  display: flex;
-  align-items: center;
-  margin-bottom: 20%;
-}
-
-.logo-img {
-  height: 80px;
-  margin-right: 10px;
-  margin-bottom: 17%;
-}
-
-.logo-text {
-  font-size: 20px;
-  font-weight: bold;
-  color: #333;
-}
-
-.nav-links {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-.nav-item {
-  text-decoration: none;
-  font-size: 16px;
-  padding: 10px;
-  background-color: #bbdefb;
-  border-radius: 5px;
-  color: #333;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  transition: background 0.3s;
-}
-
-.nav-item:hover {
-  background-color: #90caf9;
-}
-
-.icon {
-  color: #333;
-}
-
-.admin-btn {
-  margin-top: 45rem;
-}
-
-.addCalendar-btn {
-  margin-top: 27rem;
-}
-
-.manager-btn {
-  margin-top: 1rem;
-}
-
-.admin-btn button {
-  width: 100%;
-  padding: 10px;
-  background-color: #1976d2;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  transition: background 0.3s;
-  cursor: default;
-}
-
-.addCalendar-btn button {
-  width: 100%;
-  padding: 10px;
-  background-color: #1976d2;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  transition: background 0.3s;
-  cursor: default;
-}
-
-.manager-btn button {
-  width: 100%;
-  padding: 10px;
-  background-color: #1976d2;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 10px;
-  transition: background 0.3s;
-  cursor: default;
-}
-
-.admin-btn button:hover {
-  background-color: #1565c0;
-}
-
-.addCalendar-btn button:hover {
-  background-color: #1565c0;
-}
-
-.manager-btn button:hover {
-  background-color: #1565c0;
-}
