+++ frontend/src/styles/Admin.css
@@ -1,0 +1,245 @@
+.admin-container {
+    display: flex;
+    width: 100vw;
+    height: 100vh;
+    background: #fff;
+    overflow: hidden;
+    color: black;
+}
+/* alterar o nome e local pois esta a ser usado para outras pages  */
+.main-content {
+    margin-top: 1.8%;
+    margin-left: 300px;
+    padding: 20px;
+    flex-grow: 1;
+}
+
+.heading-container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+  
+.heading {
+    font-size: 24px;
+    font-weight: bold;
+    margin-top: 1%;
+}
+  
+.dropdown-container {
+    position: relative;
+}
+  
+.order-button {
+    background-color: #28a745;
+    border: none;
+    color: white;
+    padding: 8px 16px;
+    font-size: 14px;
+    cursor: pointer;
+    border-radius: 5px;
+    margin-right: 150px;
+}
+  
+.order-button:hover {
+    background-color: #218838;
+}
+  
+.dropdown-menu {
+    position: absolute;
+    top: 100%;
+    right: 0;
+    background-color: white;
+    border: 1px solid #ccc;
+    box-shadow: 0px 8px 16px rgba(0, 0, 0, 0.2);
+    z-index: 1000;
+    min-width: 150px;
+    margin-top: 5px;
+}
+  
+.dropdown-item {
+    padding: 10px;
+    cursor: pointer;
+    font-size: 14px;
+}
+  
+.dropdown-item:hover {
+    background-color: #f1f1f1;
+}
+  
+.dropdown-item:active {
+    background-color: #ddd;
+}
+
+/* Cards */
+
+.algorithm-cards {
+    display: flex;
+    justify-content: flex-start;
+    gap: 40px;
+    margin-top: 20px;
+  }
+  
+  .algorithm-card {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 220px;
+    height: 120px;
+    border-radius: 10px;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+  }
+  
+  .add-button {
+    background: transparent;
+    border: 2px solid black;
+    color: black;
+    font-size: 16px;
+    cursor: pointer;
+    margin-top: 20px;
+    text-align: left;
+  }
+  
+  .purple { background-color: #e0d4fc; }
+  .green { background-color: #d4fcd4; }
+  .yellow { background-color: #fcf4d4; }
+
+
+
+  
+
+/* slidebar admin */
+.sidebar {
+    height: 100%;
+    width: 250px;
+    background-color: #e3f2fd;
+    display: flex;
+    flex-direction: column;
+    padding: 20px;
+    position: fixed;
+    top: 0;
+    left: 0;
+    overflow: hidden;
+  }
+  
+  
+  .logo {
+    display: flex;
+    align-items: center;
+    margin-bottom: 20%;
+  }
+  
+  .logo-img {
+    height: 80px;
+    margin-right: 10px;
+    margin-bottom: 17%;
+  }
+  
+  .logo-text {
+    font-size: 20px;
+    font-weight: bold;
+    color: #333;
+  }
+  
+  .nav-links {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+  }
+  
+  .nav-item {
+    text-decoration: none;
+    font-size: 16px;
+    padding: 10px;
+    background-color: #bbdefb;
+    border-radius: 5px;
+    color: #333;
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    transition: background 0.3s;
+  }
+  
+  .nav-item:hover {
+    background-color: #90caf9;
+  }
+  
+  .icon {
+    color: #333;
+  }
+  
+  .admin-btn {
+    margin-top: 45rem;
+  }
+  
+  .addCalendar-btn {
+    margin-top: 27rem;
+  }
+  
+  .manager-btn {
+    margin-top: 1rem;
+  }
+  
+  .admin-btn button {
+    width: 100%;
+    padding: 10px;
+    background-color: #1976d2;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 16px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 10px;
+    transition: background 0.3s;
+    cursor: default;
+  }
+  
+  .addCalendar-btn button {
+    width: 100%;
+    padding: 10px;
+    background-color: #1976d2;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 16px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 10px;
+    transition: background 0.3s;
+    cursor: default;
+  }
+  
+  .manager-btn button {
+    width: 100%;
+    padding: 10px;
+    background-color: #1976d2;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 16px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 10px;
+    transition: background 0.3s;
+    cursor: default;
+  }
+  
+  .admin-btn button:hover {
+    background-color: #1565c0;
+  }
+  
+  .addCalendar-btn button:hover {
+    background-color: #1565c0;
+  }
+  
+  .manager-btn button:hover {
+    background-color: #1565c0;
+  }
