+++ frontend/src/styles/Login.css
@@ -1,0 +1,67 @@
+.login-container {
+    width: 100vw;
+    height: 100vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    font-family: sans-serif;
+  }
+  
+  .login-container::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: url('../assets/images/office-614273.jpg') center/cover no-repeat;
+    filter: blur(8px); 
+    z-index: -1;       
+  }
+  
+  .login-card {
+    position: relative;
+    background: rgba(255, 255, 255, 0.9); 
+    padding: 2rem;
+    border-radius: 16px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    width: 350px;
+    text-align: center;
+  }
+  
+  .login-logo {
+    max-width: 300px;
+    margin-bottom: 1rem;
+  }
+  
+  .login-logo img {
+    width: 500px;
+    margin-bottom: 1rem;
+  }
+  
+  .login-input {
+    width: 93%;
+    padding: 10px;
+    margin-bottom: 1rem;
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    font-size: 1rem;
+  }
+  
+  .login-button {
+    width: 50%;
+    background: #007BFF;
+    color: white;
+    padding: 10px;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+    transition: 0.3s;
+    font-size: 1rem;
+    font-weight: bold;
+  }
+  
+  .login-button:hover {
+    background: #0056b3;
+  }
